requirements.txt
sentence-transformers 
streamlit 
pandas
recsys_metrics
tqdm
torchmetrics
